feat(settings): tile-based Settings with per-category sub-pages (#801)
The Settings page mixed navigable tiles (Nozzles, Printers, …) with a
long column of inline settings that were hard to find. Make EVERY
category a tile that opens its own sub-page, like the Locations tile:

- /settings is now a pure tile grid — Manage (entities) + Configure
  (UI Settings, Network, Backup & Data, AI, Devices) + a red Danger Zone
  tile. Network is Electron-only (hidden in web).
- New sub-pages, each with a '← Settings' breadcrumb + the controls
  moved out of the monolith:
  /settings/ui (currency, theme, language), /settings/network
  (connection mode + LAN share, desktop-only), /settings/backup
  (snapshots), /settings/ai (provider + key), /settings/devices
  (NFC tools + label printer + label format), /settings/danger
  (delete all data).
- New reusable SettingsTile; ThemeSection extracted to its own component;
  LabelPrinterSettings/LabelFormatEditor wrappers de-chromed so they nest
  in cards.
- i18n: settings.group.* titles + descriptions, settings.back,
  settings.network.desktopOnly (EN/DE).

Verified in the desktop app: every tile opens its sub-page and the
breadcrumb returns to /settings.
