+++ src/app/settings/page.tsx
@@ -1,1246 +1,52 @@
 "use client";
 
-import Link from "next/link";
-import { useRef, useState, useEffect } from "react";
-import { CURRENCIES, useCurrency } from "@/hooks/useCurrency";
-import type { ValidationError } from "@/lib/customCurrency";
 import { useTranslation } from "@/i18n/TranslationProvider";
-import { useConfirm } from "@/components/ConfirmDialog";
-import { LOCALES } from "@/i18n";
-import { useTheme, type ThemePreference } from "@/components/ThemeProvider";
 import { useIsElectron } from "@/hooks/useIsElectron";
-import { useNfcContext } from "@/components/NfcProvider";
-import LabelPrinterSettings from "@/components/LabelPrinterSettings";
-import LabelFormatEditor from "@/components/LabelFormatEditor";
+import SettingsTile from "@/components/SettingsTile";
 
 export default function SettingsPage() {
-  const { t, locale, setLocale } = useTranslation();
-  const confirm = useConfirm();
-  const { notifyTagErased } = useNfcContext();
-  const [restoring, setRestoring] = useState(false);
-  const [restoreResult, setRestoreResult] = useState<{ ok: boolean; message: string } | null>(null);
-  const fileInputRef = useRef<HTMLInputElement>(null);
-
-  // Delete database state
-  const [showDeleteConfirm, setShowDeleteConfirm] = useState(false);
-  const [deleteInput, setDeleteInput] = useState("");
-  const [deleting, setDeleting] = useState(false);
-  const [deleteResult, setDeleteResult] = useState<{ ok: boolean; message: string } | null>(null);
-
-  // AI provider state
-  type AiProvider = "gemini" | "claude" | "openai";
-  const AI_PROVIDERS: { id: AiProvider; name: string; keyUrl: string }[] = [
-    { id: "gemini", name: "Google Gemini", keyUrl: "https://aistudio.google.com/apikey" },
-    { id: "claude", name: "Anthropic Claude", keyUrl: "https://console.anthropic.com/settings/keys" },
-    { id: "openai", name: "OpenAI ChatGPT", keyUrl: "https://platform.openai.com/api-keys" },
-  ];
-  const [aiProvider, setAiProvider] = useState<AiProvider>("gemini");
-  const [aiKey, setAiKey] = useState("");
-  const [aiConfigured, setAiConfigured] = useState(false);
-  const [aiSaving, setAiSaving] = useState(false);
-  const [aiResult, setAiResult] = useState<{ ok: boolean; message: string } | null>(null);
-  const [showAiKey, setShowAiKey] = useState(false);
-
-  // Currency
-  const { currency, setCurrency, customCurrencies, addCustom, removeCustom } = useCurrency();
-  const [showAddCurrency, setShowAddCurrency] = useState(false);
-  const [newCurrCode, setNewCurrCode] = useState("");
-  const [newCurrSymbol, setNewCurrSymbol] = useState("");
-  const [newCurrName, setNewCurrName] = useState("");
-  const [newCurrError, setNewCurrError] = useState<ValidationError | null>(null);
-
-  function currencyValidationMessage(err: ValidationError): string {
-    switch (err) {
-      case "code-empty":
-        return t("settings.customCurrency.error.codeEmpty");
-      case "code-too-short":
-        return t("settings.customCurrency.error.codeTooShort");
-      case "code-too-long":
-        return t("settings.customCurrency.error.codeTooLong");
-      case "code-invalid-chars":
-        return t("settings.customCurrency.error.codeInvalidChars");
-      case "code-collides-builtin":
-        return t("settings.customCurrency.error.codeCollidesBuiltin");
-      case "code-duplicate":
-        return t("settings.customCurrency.error.codeDuplicate");
-      case "symbol-empty":
-        return t("settings.customCurrency.error.symbolEmpty");
-      case "symbol-too-long":
-        return t("settings.customCurrency.error.symbolTooLong");
-    }
-  }
-
-  function handleAddCustomCurrency() {
-    const err = addCustom({ code: newCurrCode, symbol: newCurrSymbol, name: newCurrName });
-    if (err) {
-      setNewCurrError(err);
-      return;
-    }
-    // `addCustom` auto-selects the new currency on success — calling
-    // `setCurrency` here additionally would race React's state batching
-    // and reject the new code on the first click (Codex P2 follow-up
-    // to PR #142). Just reset the form.
-    setNewCurrError(null);
-    setNewCurrCode("");
-    setNewCurrSymbol("");
-    setNewCurrName("");
-    setShowAddCurrency(false);
-  }
-
-  // Connection mode state (Electron only)
-  type ConnectionMode = "atlas" | "hybrid" | "offline" | "";
-  const CONNECTION_MODES: { id: ConnectionMode; label: string; icon: string; description: string; needsUri: boolean }[] = [
-    { id: "atlas", label: t("settings.connectionAtlas"), icon: "\u2601", description: t("settings.connectionAtlasDesc"), needsUri: true },
-    { id: "hybrid", label: t("settings.connectionHybrid"), icon: "\u21C4", description: t("settings.connectionHybridDesc"), needsUri: true },
-    { id: "offline", label: t("settings.connectionOffline"), icon: "\uD83D\uDCBE", description: t("settings.connectionOfflineDesc"), needsUri: false },
-  ];
-  const [currentMode, setCurrentMode] = useState<ConnectionMode>("");
-  const [pendingMode, setPendingMode] = useState<ConnectionMode>("");
-  const [atlasUri, setAtlasUri] = useState("");
-  const [hasStoredUri, setHasStoredUri] = useState(false);
-  const [modeSwitching, setModeSwitching] = useState(false);
-  const [modeResult, setModeResult] = useState<{ ok: boolean; message: string } | null>(null);
-  const [showUriInput, setShowUriInput] = useState(false);
-  const [testingConnection, setTestingConnection] = useState(false);
-
-  // "Share on local network" (Electron only)
-  const [exposeToLan, setExposeToLan] = useState(false);
-  const [lanInfo, setLanInfo] = useState<{ ips: string[]; port: number } | null>(null);
-  const [lanSaving, setLanSaving] = useState(false);
-  const [lanError, setLanError] = useState(false);
-
-  // NFC state (Electron only)
+  const { t } = useTranslation();
   const isElectron = useIsElectron();
-  const [nfcStatus, setNfcStatus] = useState<{
-    readerConnected: boolean;
-    readerName: string | null;
-    tagPresent: boolean;
-    tagUid: string | null;
-  }>({ readerConnected: false, readerName: null, tagPresent: false, tagUid: null });
-  const [formatting, setFormatting] = useState(false);
-  const [formatResult, setFormatResult] = useState<{ ok: boolean; message: string } | null>(null);
-  const [showFormatConfirm, setShowFormatConfirm] = useState(false);
-  const [settingReadOnly, setSettingReadOnly] = useState(false);
-
-  // Load config
-  useEffect(() => {
-    const controller = new AbortController();
-    const api = window.electronAPI;
-    if (api?.getConfig) {
-      // Electron mode — check electron-store
-      api.getConfig().then((cfg) => {
-        if (controller.signal.aborted) return;
-        if (cfg.aiApiKey || cfg.geminiApiKey) {
-          setAiConfigured(true);
-        }
-        if (cfg.aiProvider) {
-          setAiProvider(cfg.aiProvider as AiProvider);
-        }
-        if (cfg.connectionMode) {
-          setCurrentMode(cfg.connectionMode as ConnectionMode);
-        }
-        if (cfg.atlasUri) {
-          setHasStoredUri(true);
-        }
-        if (typeof cfg.exposeToLan === "boolean") {
-          setExposeToLan(cfg.exposeToLan);
-        }
-      }).catch(() => {});
-      // LAN addresses to display in the "Share on local network" section.
-      api.getLanInfo?.().then((info) => {
-        if (controller.signal.aborted) return;
-        setLanInfo(info);
-      }).catch(() => {});
-    } else {
-      // Web mode — check API
-      fetch("/api/tds", { signal: controller.signal }).then((r) => r.json()).then((d) => {
-        setAiConfigured(d.configured);
-        if (d.provider) setAiProvider(d.provider);
-      }).catch(() => {});
-    }
-    return () => { controller.abort(); };
-  }, []);
-
-  // Toggle "Share on local network". Persisting flips the embedded server's
-  // bind address; saveConfig resolves only once the server has restarted, so
-  // the button stays in its "applying" state until LAN access is really live.
-  async function toggleLanShare(next: boolean) {
-    const api = window.electronAPI;
-    if (!api?.saveConfig) return;
-    setLanSaving(true);
-    setLanError(false);
-    try {
-      const res = await api.saveConfig({ exposeToLan: next });
-      if (res?.success) {
-        setExposeToLan(next);
-      } else {
-        setLanError(true);
-      }
-    } catch {
-      setLanError(true);
-    } finally {
-      setLanSaving(false);
-    }
-  }
-
-  useEffect(() => {
-    const controller = new AbortController();
-    const api = window.electronAPI;
-    if (!api?.nfcGetStatus) return;
-    api.nfcGetStatus().then((s) => {
-      if (!controller.signal.aborted) setNfcStatus(s);
-    }).catch(() => {});
-    const unsub = api.onNfcStatusChange(setNfcStatus);
-    return () => { controller.abort(); unsub(); };
-  }, []);
-
-  // Hide the erase confirmation whenever the tag isn't present.
-  // Derived in render rather than via effect so a cascading setState isn't needed.
-  const showFormatConfirmVisible = showFormatConfirm && nfcStatus.tagPresent;
-
-  const handleFormat = async () => {
-    setShowFormatConfirm(false);
-    setFormatting(true);
-    setFormatResult(null);
-    try {
-      if (!window.electronAPI?.nfcFormatTag) {
-        throw new Error("NFC format not available — restart the app to load updated NFC support");
-      }
-      await window.electronAPI.nfcFormatTag();
-      // The tag is now blank — clear the status pill's "Loaded: <name>"
-      // label immediately instead of leaving it stale until the user
-      // lifts the erased tag off the reader.
-      notifyTagErased();
-      setFormatResult({ ok: true, message: t("settings.nfcEraseSuccess") });
-    } catch (err: unknown) {
-      const raw = err instanceof Error ? err.message : String(err);
-      // GH #583: a Bambu Lab tag is read-only — surface a friendly message
-      // instead of the raw "Read block 0 failed: SW=6a81" the format throws.
-      const message = raw.includes("BAMBU_READ_ONLY")
-        ? t("settings.nfcEraseBambuReadOnly")
-        : raw;
-      setFormatResult({ ok: false, message });
-    } finally {
-      setFormatting(false);
-    }
-  };
-
-  // GH #583: toggle the soft read-only flag on the tag (reversible — CC byte
-  // write-access bits; cleared by Erase or by "Make Writable").
-  const handleSetReadOnly = async (readOnly: boolean) => {
-    setSettingReadOnly(true);
-    setFormatResult(null);
-    try {
-      if (!window.electronAPI?.nfcSetReadOnly) {
-        throw new Error("NFC read-only not available — restart the app to load updated NFC support");
-      }
-      await window.electronAPI.nfcSetReadOnly(readOnly);
-      setFormatResult({
-        ok: true,
-        message: t(readOnly ? "settings.nfcReadOnlySet" : "settings.nfcWritableSet"),
-      });
-    } catch (err: unknown) {
-      const raw = err instanceof Error ? err.message : String(err);
-      let message = raw;
-      if (raw.includes("BAMBU_READ_ONLY")) message = t("settings.nfcReadOnlyBambu");
-      else if (raw.includes("TAG_NOT_FORMATTED")) message = t("settings.nfcReadOnlyNotFormatted");
-      setFormatResult({ ok: false, message });
-    } finally {
-      setSettingReadOnly(false);
-    }
-  };
-
-  const handleRestore = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-
-    // Reset file input so the same file can be re-selected
-    e.target.value = "";
-
-    if (!(await confirm({ message: t("settings.restoreConfirm"), destructive: true, confirmLabel: t("common.restore") }))) return;
-
-    setRestoring(true);
-    setRestoreResult(null);
-
-    try {
-      const formData = new FormData();
-      formData.append("file", file);
-
-      const res = await fetch("/api/snapshot", { method: "POST", body: formData });
-      const data = await res.json();
-
-      if (res.ok) {
-        setRestoreResult({
-          ok: true,
-          message: t("settings.restoreSuccess", { filaments: data.restored.filaments, nozzles: data.restored.nozzles, printers: data.restored.printers }),
-        });
-      } else {
-        setRestoreResult({ ok: false, message: data.error || t("settings.restoreFailed") });
-      }
-    } catch {
-      setRestoreResult({ ok: false, message: t("common.serverError") });
-    } finally {
-      setRestoring(false);
-    }
-  };
 
   return (
     <main id="main-content" className="max-w-3xl mx-auto px-4 py-8">
       <h1 className="text-3xl font-bold mb-2">{t("settings.title")}</h1>
-      <p className="text-gray-500 text-sm mb-8">
-        {t("settings.subtitle")}
-      </p>
+      <p className="text-gray-500 text-sm mb-8">{t("settings.subtitle")}</p>
 
+      {/* Manage — entity catalogs, each its own page */}
+      <h2 className="text-xs font-semibold uppercase tracking-wider text-gray-500 dark:text-gray-400 mb-3">
+        {t("settings.group.manage")}
+      </h2>
       <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
-        <Link
-          href="/nozzles"
-          className="block p-5 rounded-lg border border-gray-300 dark:border-gray-700 hover:border-gray-400 dark:hover:border-gray-500 hover:bg-gray-100 dark:hover:bg-gray-900/50 transition-colors group"
-        >
-          <div className="flex items-center justify-between">
-            <div>
-              <h2 className="text-lg font-semibold text-gray-900 dark:text-gray-200 group-hover:text-black dark:group-hover:text-white">
-                {t("settings.nozzles")}
-              </h2>
-              <p className="text-sm text-gray-500 mt-1">
-                {t("settings.nozzlesDesc")}
-              </p>
-            </div>
-            <svg className="w-5 h-5 text-gray-600 group-hover:text-gray-400" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-              <path strokeLinecap="round" strokeLinejoin="round" d="M8.25 4.5l7.5 7.5-7.5 7.5" />
-            </svg>
-          </div>
-        </Link>
-
-        <Link
-          href="/printers"
-          className="block p-5 rounded-lg border border-gray-300 dark:border-gray-700 hover:border-gray-400 dark:hover:border-gray-500 hover:bg-gray-100 dark:hover:bg-gray-900/50 transition-colors group"
-        >
-          <div className="flex items-center justify-between">
-            <div>
-              <h2 className="text-lg font-semibold text-gray-900 dark:text-gray-200 group-hover:text-black dark:group-hover:text-white">
-                {t("settings.printers")}
-              </h2>
-              <p className="text-sm text-gray-500 mt-1">
-                {t("settings.printersDesc")}
-              </p>
-            </div>
-            <svg className="w-5 h-5 text-gray-600 group-hover:text-gray-400" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-              <path strokeLinecap="round" strokeLinejoin="round" d="M8.25 4.5l7.5 7.5-7.5 7.5" />
-            </svg>
-          </div>
-        </Link>
-
-        <Link
-          href="/bed-types"
-          className="block p-5 rounded-lg border border-gray-300 dark:border-gray-700 hover:border-gray-400 dark:hover:border-gray-500 hover:bg-gray-100 dark:hover:bg-gray-900/50 transition-colors group"
-        >
-          <div className="flex items-center justify-between">
-            <div>
-              <h2 className="text-lg font-semibold text-gray-900 dark:text-gray-200 group-hover:text-black dark:group-hover:text-white">
-                {t("settings.bedTypes")}
-              </h2>
-              <p className="text-sm text-gray-500 mt-1">
-                {t("settings.bedTypesDesc")}
-              </p>
-            </div>
-            <svg className="w-5 h-5 text-gray-600 group-hover:text-gray-400" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-              <path strokeLinecap="round" strokeLinejoin="round" d="M8.25 4.5l7.5 7.5-7.5 7.5" />
-            </svg>
-          </div>
-        </Link>
-
-        <Link
-          href="/locations"
-          className="block p-5 rounded-lg border border-gray-300 dark:border-gray-700 hover:border-gray-400 dark:hover:border-gray-500 hover:bg-gray-100 dark:hover:bg-gray-900/50 transition-colors group"
-        >
-          <div className="flex items-center justify-between">
-            <div>
-              <h2 className="text-lg font-semibold text-gray-900 dark:text-gray-200 group-hover:text-black dark:group-hover:text-white">
-                {t("settings.locations")}
-              </h2>
-              <p className="text-sm text-gray-500 mt-1">
-                {t("settings.locationsDesc")}
-              </p>
-            </div>
-            <svg className="w-5 h-5 text-gray-600 group-hover:text-gray-400" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-              <path strokeLinecap="round" strokeLinejoin="round" d="M8.25 4.5l7.5 7.5-7.5 7.5" />
-            </svg>
-          </div>
-        </Link>
-
-        <Link
-          href="/trash"
-          className="block p-5 rounded-lg border border-gray-300 dark:border-gray-700 hover:border-gray-400 dark:hover:border-gray-500 hover:bg-gray-100 dark:hover:bg-gray-900/50 transition-colors group"
-        >
-          <div className="flex items-center justify-between">
-            <div>
-              <h2 className="text-lg font-semibold text-gray-900 dark:text-gray-200 group-hover:text-black dark:group-hover:text-white">
-                {t("settings.trash")}
-              </h2>
-              <p className="text-sm text-gray-500 mt-1">
-                {t("settings.trashDesc")}
-              </p>
-            </div>
-            <svg className="w-5 h-5 text-gray-600 group-hover:text-gray-400" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-              <path strokeLinecap="round" strokeLinejoin="round" d="M8.25 4.5l7.5 7.5-7.5 7.5" />
-            </svg>
-          </div>
-        </Link>
-
-        <Link
-          href="/import-export"
-          className="block p-5 rounded-lg border border-gray-300 dark:border-gray-700 hover:border-gray-400 dark:hover:border-gray-500 hover:bg-gray-100 dark:hover:bg-gray-900/50 transition-colors group"
-        >
-          <div className="flex items-center justify-between">
-            <div>
-              <h2 className="text-lg font-semibold text-gray-900 dark:text-gray-200 group-hover:text-black dark:group-hover:text-white">
-                {t("settings.importExport")}
-              </h2>
-              <p className="text-sm text-gray-500 mt-1">
-                {t("settings.importExportDesc")}
-              </p>
-            </div>
-            <svg className="w-5 h-5 text-gray-600 group-hover:text-gray-400" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-              <path strokeLinecap="round" strokeLinejoin="round" d="M8.25 4.5l7.5 7.5-7.5 7.5" />
-            </svg>
-          </div>
-        </Link>
-
-        <Link
-          href="/api-docs"
-          className="block p-5 rounded-lg border border-gray-300 dark:border-gray-700 hover:border-gray-400 dark:hover:border-gray-500 hover:bg-gray-100 dark:hover:bg-gray-900/50 transition-colors group"
-        >
-          <div className="flex items-center justify-between">
-            <div>
-              <h2 className="text-lg font-semibold text-gray-900 dark:text-gray-200 group-hover:text-black dark:group-hover:text-white">
-                {t("settings.apiDocs")}
-              </h2>
-              <p className="text-sm text-gray-500 mt-1">
-                {t("settings.apiDocsDesc")}
-              </p>
-            </div>
-            <svg className="w-5 h-5 text-gray-600 group-hover:text-gray-400" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-              <path strokeLinecap="round" strokeLinejoin="round" d="M8.25 4.5l7.5 7.5-7.5 7.5" />
-            </svg>
-          </div>
-        </Link>
+        <SettingsTile href="/nozzles" title={t("settings.nozzles")} description={t("settings.nozzlesDesc")} />
+        <SettingsTile href="/printers" title={t("settings.printers")} description={t("settings.printersDesc")} />
+        <SettingsTile href="/bed-types" title={t("settings.bedTypes")} description={t("settings.bedTypesDesc")} />
+        <SettingsTile href="/locations" title={t("settings.locations")} description={t("settings.locationsDesc")} />
+        <SettingsTile href="/trash" title={t("settings.trash")} description={t("settings.trashDesc")} />
+        <SettingsTile href="/import-export" title={t("settings.importExport")} description={t("settings.importExportDesc")} />
+        <SettingsTile href="/api-docs" title={t("settings.apiDocs")} description={t("settings.apiDocsDesc")} />
       </div>
 
-      {/* Currency */}
-      <div className="mt-8 pt-6 border-t border-gray-200 dark:border-gray-800">
-        <h2 className="text-lg font-semibold text-gray-900 dark:text-gray-200 mb-1">{t("settings.currency")}</h2>
-        <p className="text-sm text-gray-500 mb-4">
-          {t("settings.currencyDesc")}
-        </p>
-        <div className="flex flex-wrap gap-2">
-          {CURRENCIES.map((c) => (
-            <button
-              key={c.code}
-              type="button"
-              onClick={() => setCurrency(c.code)}
-              className={`px-4 py-2 text-sm rounded border transition-colors ${
-                currency === c.code
-                  ? "border-blue-500 bg-blue-50 dark:bg-blue-600/20 text-blue-600 dark:text-blue-300"
-                  : "border-gray-300 dark:border-gray-700 text-gray-600 dark:text-gray-400 hover:border-gray-400 dark:hover:border-gray-500 hover:text-gray-800 dark:hover:text-gray-300"
-              }`}
-              title={c.name}
-            >
-              {c.symbol} {c.code}
-            </button>
-          ))}
-          {customCurrencies.map((c) => (
-            <span
-              key={c.code}
-              className={`inline-flex items-center text-sm rounded border transition-colors ${
-                currency === c.code
-                  ? "border-blue-500 bg-blue-50 dark:bg-blue-600/20 text-blue-600 dark:text-blue-300"
-                  : "border-gray-300 dark:border-gray-700 text-gray-600 dark:text-gray-400 hover:border-gray-400 dark:hover:border-gray-500"
-              }`}
-            >
-              <button
-                type="button"
-                onClick={() => setCurrency(c.code)}
-                className="px-3 py-2"
-                title={c.name || c.code}
-              >
-                {c.symbol} {c.code}
-              </button>
-              <button
-                type="button"
-                onClick={() => removeCustom(c.code)}
-                className="px-2 py-2 border-l border-current opacity-60 hover:opacity-100"
-                aria-label={t("settings.customCurrency.removeAria", { code: c.code })}
-                title={t("settings.customCurrency.remove")}
-              >
-                ×
-              </button>
-            </span>
-          ))}
-          <button
-            type="button"
-            onClick={() => setShowAddCurrency((v) => !v)}
-            className="px-4 py-2 text-sm rounded border border-dashed border-gray-300 dark:border-gray-700 text-gray-600 dark:text-gray-400 hover:border-gray-400 dark:hover:border-gray-500 hover:text-gray-800 dark:hover:text-gray-300"
-          >
-            {showAddCurrency ? t("settings.customCurrency.cancel") : t("settings.customCurrency.add")}
-          </button>
-        </div>
-
-        {showAddCurrency && (
-          <div className="mt-4 p-4 rounded border border-gray-200 dark:border-gray-700 bg-gray-50 dark:bg-gray-900/50">
-            <div className="grid grid-cols-1 sm:grid-cols-3 gap-3">
-              <div>
-                <label htmlFor="custom-currency-code" className="block text-xs font-semibold text-gray-500 uppercase tracking-wider mb-1">
-                  {t("settings.customCurrency.codeLabel")}
-                </label>
-                <input
-                  id="custom-currency-code"
-                  type="text"
-                  value={newCurrCode}
-                  onChange={(e) => {
-                    setNewCurrCode(e.target.value);
-                    setNewCurrError(null);
-                  }}
-                  placeholder="SEK"
-                  maxLength={6}
-                  className="w-full px-3 py-2 text-sm rounded border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-200 uppercase"
-                />
-              </div>
-              <div>
-                <label htmlFor="custom-currency-symbol" className="block text-xs font-semibold text-gray-500 uppercase tracking-wider mb-1">
-                  {t("settings.customCurrency.symbolLabel")}
-                </label>
-                <input
-                  id="custom-currency-symbol"
-                  type="text"
-                  value={newCurrSymbol}
-                  onChange={(e) => {
-                    setNewCurrSymbol(e.target.value);
-                    setNewCurrError(null);
-                  }}
-                  placeholder="kr"
-                  maxLength={4}
-                  className="w-full px-3 py-2 text-sm rounded border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-200"
-                />
-              </div>
-              <div>
-                <label htmlFor="custom-currency-name" className="block text-xs font-semibold text-gray-500 uppercase tracking-wider mb-1">
-                  {t("settings.customCurrency.nameLabel")}
-                </label>
-                <input
-                  id="custom-currency-name"
-                  type="text"
-                  value={newCurrName}
-                  onChange={(e) => setNewCurrName(e.target.value)}
-                  placeholder={t("settings.customCurrency.namePlaceholder")}
-                  className="w-full px-3 py-2 text-sm rounded border border-gray-300 dark:border-gray-600 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-200"
-                />
-              </div>
-            </div>
-            {newCurrError && (
-              <p className="mt-2 text-xs text-red-600 dark:text-red-400">
-                {currencyValidationMessage(newCurrError)}
-              </p>
-            )}
-            <div className="mt-3 flex gap-2">
-              <button
-                type="button"
-                onClick={handleAddCustomCurrency}
-                className="px-4 py-2 text-sm rounded bg-blue-600 hover:bg-blue-700 text-white"
-              >
-                {t("settings.customCurrency.save")}
-              </button>
-            </div>
-          </div>
+      {/* Configure — preferences + integrations, each its own page */}
+      <h2 className="text-xs font-semibold uppercase tracking-wider text-gray-500 dark:text-gray-400 mb-3 mt-10">
+        {t("settings.group.configure")}
+      </h2>
+      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
+        <SettingsTile href="/settings/ui" title={t("settings.group.ui")} description={t("settings.group.ui.desc")} />
+        {/* Network controls (connection mode, LAN share) are desktop-only, so
+            the tile only appears in Electron. */}
+        {isElectron && (
+          <SettingsTile href="/settings/network" title={t("settings.group.network")} description={t("settings.group.network.desc")} />
         )}
+        <SettingsTile href="/settings/backup" title={t("settings.group.data")} description={t("settings.group.data.desc")} />
+        <SettingsTile href="/settings/ai" title={t("settings.group.ai")} description={t("settings.group.ai.desc")} />
+        <SettingsTile href="/settings/devices" title={t("settings.group.devices")} description={t("settings.group.devices.desc")} />
       </div>
 
-      {/* Theme */}
-      <ThemeSection />
-
-      {/* Language */}
-      <div className="mt-8 pt-6 border-t border-gray-200 dark:border-gray-800">
-        <h2 className="text-lg font-semibold text-gray-900 dark:text-gray-200 mb-1">{t("settings.language")}</h2>
-        <p className="text-sm text-gray-500 mb-4">
-          {t("settings.languageDesc")}
-        </p>
-        <div className="flex gap-2">
-          {LOCALES.map((l) => (
-            <button
-              key={l.code}
-              type="button"
-              onClick={() => setLocale(l.code)}
-              className={`px-4 py-2 text-sm rounded border transition-colors ${
-                locale === l.code
-                  ? "border-blue-500 bg-blue-50 dark:bg-blue-600/20 text-blue-600 dark:text-blue-300"
-                  : "border-gray-300 dark:border-gray-700 text-gray-600 dark:text-gray-400 hover:border-gray-400 dark:hover:border-gray-500 hover:text-gray-800 dark:hover:text-gray-300"
-              }`}
-            >
-              {l.nativeName}
-            </button>
-          ))}
-        </div>
-      </div>
-
-      {/* Connection Mode (Electron only) */}
-      {isElectron && currentMode && (
-        <div className="mt-8 pt-6 border-t border-gray-200 dark:border-gray-800">
-          <h2 className="text-lg font-semibold text-gray-900 dark:text-gray-200 mb-1">{t("settings.connectionMode")}</h2>
-          <p className="text-sm text-gray-500 mb-4">
-            {t("settings.connectionModeDesc")}
-          </p>
-
-          <div className="space-y-2">
-            {CONNECTION_MODES.map((m) => {
-              const isActive = currentMode === m.id;
-              const isPending = pendingMode === m.id;
-              return (
-                <button
-                  key={m.id}
-                  type="button"
-                  disabled={modeSwitching}
-                  onClick={() => {
-                    if (isActive) return;
-                    setModeResult(null);
-                    if (m.needsUri && !hasStoredUri) {
-                      // Need a URI — show input
-                      setPendingMode(m.id);
-                      setShowUriInput(true);
-                    } else {
-                      // Can switch directly
-                      setPendingMode(m.id);
-                      setShowUriInput(false);
-                    }
-                  }}
-                  className={`w-full text-left p-3 rounded-lg border transition-colors ${
-                    isActive
-                      ? "border-blue-500 bg-blue-600/10"
-                      : isPending
-                        ? "border-yellow-500 bg-yellow-500/5"
-                        : "border-gray-300 dark:border-gray-700 hover:border-gray-400 dark:hover:border-gray-500"
-                  } ${modeSwitching ? "opacity-50 cursor-not-allowed" : ""}`}
-                >
-                  <div className="flex items-center gap-3">
-                    <span className="text-xl">{m.icon}</span>
-                    <div className="flex-1">
-                      <div className="flex items-center gap-2">
-                        <span className={`text-sm font-medium ${isActive ? "text-blue-600 dark:text-blue-400" : "text-gray-700 dark:text-gray-200"}`}>
-                          {m.label}
-                        </span>
-                        {isActive && (
-                          <span className="text-xs px-1.5 py-0.5 bg-blue-100 dark:bg-blue-600/30 text-blue-600 dark:text-blue-300 rounded">{t("settings.connectionCurrent")}</span>
-                        )}
-                      </div>
-                      <p className="text-xs text-gray-500 mt-0.5">{m.description}</p>
-                    </div>
-                  </div>
-                </button>
-              );
-            })}
-          </div>
-
-          {/* URI input for modes that need it */}
-          {showUriInput && pendingMode && (
-            <div className="mt-3 p-4 border border-gray-300 dark:border-gray-700 rounded-lg bg-gray-50 dark:bg-gray-900/50">
-              <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-                {t("settings.atlasConnectionString")}
-              </label>
-              <input
-                type="password"
-                value={atlasUri}
-                onChange={(e) => setAtlasUri(e.target.value)}
-                placeholder="mongodb+srv://user:pass@cluster.mongodb.net/filament-db"
-                className="w-full px-3 py-2 bg-transparent border border-gray-300 dark:border-gray-700 rounded text-sm text-gray-700 dark:text-gray-200 placeholder-gray-400 dark:placeholder-gray-600 focus:outline-none focus:border-blue-600 mb-2"
-                autoFocus
-              />
-              <p className="text-xs text-gray-500 mb-3">
-                {t("settings.connectionStringPrivacy")}
-              </p>
-              <div className="flex gap-2">
-                <button
-                  onClick={async () => {
-                    if (!atlasUri.trim()) return;
-                    setTestingConnection(true);
-                    setModeResult(null);
-                    try {
-                      const result = await window.electronAPI!.testConnection(atlasUri.trim());
-                      if (!result.success) {
-                        setModeResult({ ok: false, message: result.error || t("settings.connectionFailed") });
-                        setTestingConnection(false);
-                        return;
-                      }
-                      // Connection OK — save and switch
-                      setModeSwitching(true);
-                      await window.electronAPI!.saveConfig({
-                        connectionMode: pendingMode,
-                        atlasUri: atlasUri.trim(),
-                      });
-                      setCurrentMode(pendingMode);
-                      setHasStoredUri(true);
-                      setPendingMode("");
-                      setShowUriInput(false);
-                      setAtlasUri("");
-                      setModeResult({ ok: true, message: t("settings.switchedTo", { mode: CONNECTION_MODES.find((m) => m.id === pendingMode)?.label || "" }) });
-                    } catch (err) {
-                      setModeResult({ ok: false, message: err instanceof Error ? err.message : t("settings.switchFailed") });
-                    } finally {
-                      setTestingConnection(false);
-                      setModeSwitching(false);
-                    }
-                  }}
-                  disabled={testingConnection || !atlasUri.trim()}
-                  className="px-4 py-1.5 bg-blue-600 text-white rounded text-sm hover:bg-blue-700 disabled:opacity-50 transition-colors"
-                >
-                  {testingConnection ? t("settings.testing") : t("settings.connectAndSwitch")}
-                </button>
-                <button
-                  onClick={() => { setPendingMode(""); setShowUriInput(false); setAtlasUri(""); setModeResult(null); }}
-                  className="px-3 py-1.5 text-gray-500 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-200 text-sm transition-colors"
-                >
-                  {t("common.cancel")}
-                </button>
-              </div>
-            </div>
-          )}
-
-          {/* Confirm button for direct switches (already have URI or switching to offline) */}
-          {pendingMode && !showUriInput && (
-            <div className="mt-3 flex items-center gap-3">
-              <button
-                onClick={async () => {
-                  setModeSwitching(true);
-                  setModeResult(null);
-                  try {
-                    await window.electronAPI!.saveConfig({ connectionMode: pendingMode });
-                    setCurrentMode(pendingMode);
-                    setPendingMode("");
-                    setModeResult({ ok: true, message: t("settings.switchedTo", { mode: CONNECTION_MODES.find((m) => m.id === pendingMode)?.label || "" }) });
-                  } catch (err) {
-                    setModeResult({ ok: false, message: err instanceof Error ? err.message : t("settings.switchFailed") });
-                  } finally {
-                    setModeSwitching(false);
-                  }
-                }}
-                disabled={modeSwitching}
-                className="px-4 py-1.5 bg-yellow-600 text-white rounded text-sm hover:bg-yellow-500 disabled:opacity-50 transition-colors"
-              >
-                {modeSwitching ? t("settings.switching") : t("settings.switchTo", { mode: CONNECTION_MODES.find((m) => m.id === pendingMode)?.label || "" })}
-              </button>
-              <button
-                onClick={() => { setPendingMode(""); setModeResult(null); }}
-                className="px-3 py-1.5 text-gray-500 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-200 text-sm transition-colors"
-              >
-                {t("common.cancel")}
-              </button>
-            </div>
-          )}
-
-          {modeResult && (
-            <div className={`mt-3 text-sm px-3 py-2 rounded ${
-              modeResult.ok
-                ? "bg-green-50 dark:bg-green-900/50 text-green-700 dark:text-green-300 border border-green-200 dark:border-green-800"
-                : "bg-red-50 dark:bg-red-900/50 text-red-700 dark:text-red-300 border border-red-200 dark:border-red-800"
-            }`}>
-              {modeResult.message}
-            </div>
-          )}
-        </div>
-      )}
-
-      {/* Share on local network (Electron only) */}
-      {isElectron && (
-        <div className="mt-8 pt-6 border-t border-gray-200 dark:border-gray-800">
-          <h2 className="text-lg font-semibold text-gray-800 dark:text-gray-200 mb-1">
-            {t("settings.lanShare.title")}
-          </h2>
-          <p className="text-sm text-gray-500 mb-4">{t("settings.lanShare.desc")}</p>
-
-          <div className="flex items-center gap-3 flex-wrap">
-            <button
-              type="button"
-              role="switch"
-              aria-checked={exposeToLan}
-              disabled={lanSaving}
-              onClick={() => toggleLanShare(!exposeToLan)}
-              className={`px-4 py-2 text-sm rounded border transition-colors disabled:opacity-50 ${
-                exposeToLan
-                  ? "border-green-500 bg-green-50 dark:bg-green-600/20 text-green-700 dark:text-green-300"
-                  : "border-gray-300 dark:border-gray-700 text-gray-600 dark:text-gray-400 hover:border-gray-400 dark:hover:border-gray-500 hover:text-gray-800 dark:hover:text-gray-300"
-              }`}
-            >
-              {lanSaving
-                ? t("settings.lanShare.applying")
-                : exposeToLan
-                  ? t("settings.lanShare.on")
-                  : t("settings.lanShare.off")}
-            </button>
-            {exposeToLan && !lanSaving && (
-              <span className="text-sm text-green-700 dark:text-green-300">
-                {t("settings.lanShare.activeHint")}
-              </span>
-            )}
-          </div>
-
-          {lanError && (
-            <div className="mt-3 text-sm px-3 py-2 rounded bg-red-50 dark:bg-red-900/50 text-red-700 dark:text-red-300 border border-red-200 dark:border-red-800">
-              {t("settings.lanShare.error")}
-            </div>
-          )}
-
-          {exposeToLan && (
-            <div className="mt-4">
-              <p className="text-sm text-gray-600 dark:text-gray-400 mb-3">
-                {t("settings.lanShare.autoDiscover")}
-              </p>
-              {lanInfo && lanInfo.ips.length > 0 ? (
-                <>
-                  <p className="text-sm text-gray-600 dark:text-gray-400 mb-2">
-                    {t("settings.lanShare.connectAt")}
-                  </p>
-                  <ul className="space-y-1">
-                    {lanInfo.ips.map((ip) => (
-                      <li key={ip}>
-                        <code className="text-sm px-2 py-1 bg-gray-100 dark:bg-gray-800 rounded font-mono text-gray-800 dark:text-gray-200">
-                          http://{ip}:{lanInfo.port}
-                        </code>
-                      </li>
-                    ))}
-                  </ul>
-                </>
-              ) : (
-                <p className="text-sm text-amber-600 dark:text-amber-400">
-                  {t("settings.lanShare.noIp")}
-                </p>
-              )}
-            </div>
-          )}
-
-          <div className="mt-4 text-xs text-amber-700 dark:text-amber-300/90 bg-amber-50 dark:bg-amber-900/20 border border-amber-200 dark:border-amber-800/60 rounded px-3 py-2">
-            {t("settings.lanShare.warning")}
-          </div>
-        </div>
-      )}
-
-      {/* Database Snapshots */}
-      <div className="mt-8 pt-6 border-t border-gray-200 dark:border-gray-800">
-        <h2 className="text-lg font-semibold text-gray-800 dark:text-gray-200 mb-1">{t("settings.snapshots")}</h2>
-        <p className="text-sm text-gray-500 mb-4">
-          {t("settings.snapshotsDesc")}
-        </p>
-
-        <div className="flex gap-3 items-center flex-wrap">
-          <a
-            href="/api/snapshot"
-            className="px-4 py-2 bg-blue-600 text-white rounded text-sm hover:bg-blue-700 transition-colors inline-flex items-center gap-2"
-          >
-            <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-              <path strokeLinecap="round" strokeLinejoin="round" d="M3 16.5v2.25A2.25 2.25 0 0 0 5.25 21h13.5A2.25 2.25 0 0 0 21 18.75V16.5M16.5 12 12 16.5m0 0L7.5 12m4.5 4.5V3" />
-            </svg>
-            {t("settings.downloadSnapshot")}
-          </a>
-
-          <input
-            ref={fileInputRef}
-            type="file"
-            accept=".json"
-            className="hidden"
-            onChange={handleRestore}
-          />
-          <button
-            onClick={() => fileInputRef.current?.click()}
-            disabled={restoring}
-            className="px-4 py-2 bg-gray-200 dark:bg-gray-700 text-gray-900 dark:text-white rounded text-sm hover:bg-gray-300 dark:hover:bg-gray-600 disabled:opacity-50 transition-colors inline-flex items-center gap-2"
-          >
-            <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-              <path strokeLinecap="round" strokeLinejoin="round" d="M3 16.5v2.25A2.25 2.25 0 0 0 5.25 21h13.5A2.25 2.25 0 0 0 21 18.75V16.5m-13.5-9L12 3m0 0 4.5 4.5M12 3v13.5" />
-            </svg>
-            {restoring ? t("settings.restoring") : t("settings.restoreFromSnapshot")}
-          </button>
-        </div>
-
-        {restoreResult && (
-          <div className={`mt-3 text-sm px-3 py-2 rounded ${
-            restoreResult.ok
-              ? "bg-green-50 dark:bg-green-900/50 text-green-700 dark:text-green-300 border border-green-200 dark:border-green-800"
-              : "bg-red-50 dark:bg-red-900/50 text-red-700 dark:text-red-300 border border-red-200 dark:border-red-800"
-          }`}>
-            {restoreResult.message}
-          </div>
-        )}
-      </div>
-
-      {/* AI Features — Provider & API Key */}
-      <div className="mt-8 pt-6 border-t border-gray-200 dark:border-gray-800">
-        <h2 className="text-lg font-semibold text-gray-800 dark:text-gray-200 mb-1">{t("settings.aiFeatures")}</h2>
-        <p className="text-sm text-gray-500 mb-4">
-          {t("settings.aiFeaturesDesc")}
-        </p>
-
-        <div className="flex items-center gap-2 mb-4">
-          <span className={`inline-block w-2.5 h-2.5 rounded-full ${aiConfigured ? "bg-green-500" : "bg-gray-600"}`} />
-          <span className="text-sm text-gray-500 dark:text-gray-400">
-            {aiConfigured
-              ? t("settings.aiConfigured", { provider: AI_PROVIDERS.find((p) => p.id === aiProvider)?.name || aiProvider })
-              : t("settings.aiNotConfigured")}
-          </span>
-        </div>
-
-        {/* Provider selector */}
-        <div className="mb-3">
-          <label className="text-xs text-gray-500 block mb-1.5 font-medium uppercase tracking-wider">{t("settings.aiProvider")}</label>
-          <div className="flex gap-2">
-            {AI_PROVIDERS.map((p) => (
-              <button
-                key={p.id}
-                type="button"
-                onClick={() => { setAiProvider(p.id); setAiResult(null); }}
-                className={`px-3 py-1.5 text-sm rounded border transition-colors ${
-                  aiProvider === p.id
-                    ? "border-blue-500 bg-blue-50 dark:bg-blue-600/20 text-blue-600 dark:text-blue-300"
-                    : "border-gray-300 dark:border-gray-700 text-gray-500 dark:text-gray-400 hover:border-gray-400 dark:hover:border-gray-500 hover:text-gray-700 dark:hover:text-gray-300"
-                }`}
-              >
-                {p.name}
-              </button>
-            ))}
-          </div>
-        </div>
-
-        {/* Cost info note */}
-        <div className="mb-3 text-xs px-3 py-2 rounded border border-amber-200 dark:border-amber-800 bg-amber-50 dark:bg-amber-900/40 text-amber-700 dark:text-amber-300">
-          {aiProvider === "gemini"
-            ? t("settings.aiCostNote.gemini")
-            : t("settings.aiCostNote.paid", { provider: AI_PROVIDERS.find((p) => p.id === aiProvider)?.name || aiProvider })}
-        </div>
-
-        {/* API key link */}
-        <p className="text-xs text-gray-500 mb-2">
-          {t("settings.getKeyFrom")}{" "}
-          <a
-            href={AI_PROVIDERS.find((p) => p.id === aiProvider)?.keyUrl}
-            target="_blank"
-            rel="noopener noreferrer"
-            className="text-blue-400 hover:text-blue-300 underline"
-          >
-            {AI_PROVIDERS.find((p) => p.id === aiProvider)?.name}
-          </a>
-        </p>
-
-        <div className="flex gap-2 items-center flex-wrap">
-          <div className="relative flex-1 max-w-sm">
-            <input
-              type={showAiKey ? "text" : "password"}
-              value={aiKey}
-              onChange={(e) => setAiKey(e.target.value)}
-              placeholder={aiConfigured ? "••••••••••••••••" : t("settings.enterApiKey")}
-              className="w-full px-3 py-2 bg-transparent border border-gray-300 dark:border-gray-700 rounded text-sm text-gray-800 dark:text-gray-200 placeholder-gray-400 dark:placeholder-gray-600 focus:outline-none focus:border-blue-600"
-            />
-            <button
-              type="button"
-              onClick={() => setShowAiKey((s) => !s)}
-              className="absolute right-2 top-1/2 -translate-y-1/2 text-gray-400 dark:text-gray-500 hover:text-gray-600 dark:hover:text-gray-300 text-xs"
-            >
-              {showAiKey ? t("common.hide") : t("common.show")}
-            </button>
-          </div>
-
-          <button
-            onClick={async () => {
-              if (!aiKey.trim()) return;
-              setAiSaving(true);
-              setAiResult(null);
-              try {
-                const api = window.electronAPI;
-                if (api?.saveConfig) {
-                  // Electron mode
-                  await api.saveConfig({ aiApiKey: aiKey.trim(), aiProvider });
-                  setAiConfigured(true);
-                  setAiKey("");
-                  setAiResult({ ok: true, message: t("settings.aiKeySaved", { provider: AI_PROVIDERS.find((p) => p.id === aiProvider)?.name || aiProvider }) });
-                } else {
-                  // Web mode — validate via API
-                  const res = await fetch("/api/tds", {
-                    method: "PUT",
-                    headers: { "Content-Type": "application/json" },
-                    body: JSON.stringify({ apiKey: aiKey.trim(), provider: aiProvider }),
-                  });
-                  const data = await res.json();
-                  if (res.ok) {
-                    setAiConfigured(true);
-                    setAiKey("");
-                    setAiResult({ ok: true, message: t("settings.aiKeySavedValidated", { provider: AI_PROVIDERS.find((p) => p.id === aiProvider)?.name || aiProvider }) });
-                  } else {
-                    setAiResult({ ok: false, message: data.error || t("settings.aiKeySaveFailed") });
-                  }
-                }
-              } catch {
-                setAiResult({ ok: false, message: t("settings.aiKeySaveFailed") });
-              } finally {
-                setAiSaving(false);
-              }
-            }}
-            disabled={aiSaving || !aiKey.trim()}
-            className="px-4 py-2 bg-blue-600 text-white rounded text-sm hover:bg-blue-700 disabled:opacity-50 transition-colors"
-          >
-            {aiSaving ? t("settings.validating") : t("settings.saveKey")}
-          </button>
-
-          {aiConfigured && (
-            <button
-              onClick={async () => {
-                const api = window.electronAPI;
-                if (api?.saveConfig) {
-                  await api.saveConfig({ aiApiKey: "", aiProvider: "gemini" });
-                } else {
-                  await fetch("/api/tds", { method: "DELETE" });
-                }
-                setAiConfigured(false);
-                // Keep the visible provider selector in sync with what we just
-                // persisted (aiProvider: "gemini"); otherwise it keeps showing
-                // the old provider after the key is removed (#683).
-                setAiProvider("gemini");
-                setAiResult({ ok: true, message: t("settings.aiKeyRemoved") });
-              }}
-              className="px-4 py-2 bg-gray-200 dark:bg-gray-700 text-gray-900 dark:text-white rounded text-sm hover:bg-gray-300 dark:hover:bg-gray-600 transition-colors"
-            >
-              {t("settings.removeKey")}
-            </button>
-          )}
-        </div>
-
-        {aiResult && (
-          <div className={`mt-3 text-sm px-3 py-2 rounded ${
-            aiResult.ok
-              ? "bg-green-50 dark:bg-green-900/50 text-green-700 dark:text-green-300 border border-green-200 dark:border-green-800"
-              : "bg-red-50 dark:bg-red-900/50 text-red-700 dark:text-red-300 border border-red-200 dark:border-red-800"
-          }`}>
-            {aiResult.message}
-          </div>
-        )}
-      </div>
-
-      {/* NFC Tools (Electron only) */}
-      {isElectron && (
-        <div className="mt-8 pt-6 border-t border-gray-200 dark:border-gray-800">
-          <h2 className="text-lg font-semibold text-gray-800 dark:text-gray-200 mb-1">{t("settings.nfcTools")}</h2>
-          <p className="text-sm text-gray-500 mb-4">
-            {t("settings.nfcToolsDesc")}
-          </p>
-
-          {/* GH #583: the reader/tag status indicator that used to live here
-              is redundant with the global NFC pill in the app header
-              (NfcStatus), so it was removed. The Erase button's disabled
-              state below still gates on nfcStatus.tagPresent. */}
-          {!showFormatConfirmVisible ? (
-            <button
-              onClick={() => { setShowFormatConfirm(true); setFormatResult(null); }}
-              disabled={formatting || !nfcStatus.tagPresent}
-              className="px-4 py-2 bg-gray-200 dark:bg-gray-700 text-gray-900 dark:text-white rounded text-sm hover:bg-gray-300 dark:hover:bg-gray-600 disabled:opacity-30 disabled:cursor-not-allowed transition-colors inline-flex items-center gap-2"
-            >
-              <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-                <path strokeLinecap="round" strokeLinejoin="round" d="M12 9.75 14.25 12m0 0 2.25 2.25M14.25 12l2.25-2.25M14.25 12 12 14.25m-2.58 4.92-6.374-6.375a1.125 1.125 0 0 1 0-1.59L9.42 4.83a1.125 1.125 0 0 1 1.59 0l6.375 6.375a1.125 1.125 0 0 1 0 1.59l-6.375 6.375a1.125 1.125 0 0 1-1.59 0Z" />
-              </svg>
-              {t("settings.eraseTag")}
-            </button>
-          ) : (
-            <div className="p-4 border border-yellow-300 dark:border-yellow-800 rounded-lg bg-yellow-50 dark:bg-yellow-950/30">
-              <p className="text-sm text-yellow-800 dark:text-yellow-300 mb-3">
-                {t("settings.eraseConfirm")}
-              </p>
-              <div className="flex gap-2 items-center">
-                <button
-                  onClick={handleFormat}
-                  disabled={formatting}
-                  className="px-4 py-1.5 bg-yellow-700 text-white rounded text-sm hover:bg-yellow-600 disabled:opacity-50 transition-colors"
-                >
-                  {formatting ? t("settings.erasing") : t("settings.confirmErase")}
-                </button>
-                <button
-                  onClick={() => setShowFormatConfirm(false)}
-                  className="px-3 py-1.5 text-gray-500 dark:text-gray-400 hover:text-gray-700 dark:hover:text-gray-200 text-sm transition-colors"
-                >
-                  {t("common.cancel")}
-                </button>
-              </div>
-            </div>
-          )}
-
-          {/* GH #583: soft read-only toggle. Reversible — "Make Writable" or
-              Erase clears it. Honored by the Write NFC flow on the filament
-              detail page. */}
-          <div className="mt-3 flex flex-wrap items-center gap-2">
-            <button
-              onClick={() => handleSetReadOnly(true)}
-              disabled={settingReadOnly || formatting || !nfcStatus.tagPresent}
-              className="px-3 py-1.5 bg-gray-200 dark:bg-gray-700 text-gray-900 dark:text-white rounded text-sm hover:bg-gray-300 dark:hover:bg-gray-600 disabled:opacity-30 disabled:cursor-not-allowed transition-colors inline-flex items-center gap-2"
-            >
-              <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-                <path strokeLinecap="round" strokeLinejoin="round" d="M16.5 10.5V6.75a4.5 4.5 0 1 0-9 0v3.75m-.75 11.25h10.5a2.25 2.25 0 0 0 2.25-2.25v-6.75a2.25 2.25 0 0 0-2.25-2.25H6.75a2.25 2.25 0 0 0-2.25 2.25v6.75a2.25 2.25 0 0 0 2.25 2.25Z" />
-              </svg>
-              {settingReadOnly ? t("settings.nfcReadOnlyWorking") : t("settings.nfcSetReadOnly")}
-            </button>
-            <button
-              onClick={() => handleSetReadOnly(false)}
-              disabled={settingReadOnly || formatting || !nfcStatus.tagPresent}
-              className="px-3 py-1.5 text-gray-600 dark:text-gray-300 rounded text-sm hover:bg-gray-100 dark:hover:bg-gray-800 disabled:opacity-30 disabled:cursor-not-allowed transition-colors inline-flex items-center gap-2"
-            >
-              <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-                <path strokeLinecap="round" strokeLinejoin="round" d="M13.5 10.5V6.75a4.5 4.5 0 1 1 9 0v3.75M3.75 21.75h10.5a2.25 2.25 0 0 0 2.25-2.25v-6.75a2.25 2.25 0 0 0-2.25-2.25H3.75a2.25 2.25 0 0 0-2.25 2.25v6.75a2.25 2.25 0 0 0 2.25 2.25Z" />
-              </svg>
-              {t("settings.nfcMakeWritable")}
-            </button>
-          </div>
-          <p className="mt-2 text-xs text-gray-400 dark:text-gray-500">
-            {t("settings.nfcReadOnlyHint")}
-          </p>
-
-          {formatResult && (
-            <div className={`mt-3 text-sm px-3 py-2 rounded ${
-              formatResult.ok
-                ? "bg-green-50 dark:bg-green-900/50 text-green-700 dark:text-green-300 border border-green-200 dark:border-green-800"
-                : "bg-red-50 dark:bg-red-900/50 text-red-700 dark:text-red-300 border border-red-200 dark:border-red-800"
-            }`}>
-              {formatResult.message}
-            </div>
-          )}
-        </div>
-      )}
-
-      {/* Brother PT-P710BT label printer device picker. Self-gates on
-          isElectron so it disappears entirely in web mode. The label format
-          editor (#592) is NOT gated — the layout also applies to the web
-          .bin-download path, so web users can configure it too. */}
-      <div className="mt-8 pt-6 border-t border-gray-200 dark:border-gray-800">
-        <LabelPrinterSettings />
-        <LabelFormatEditor />
-      </div>
-
-      {/* Danger Zone — kept LAST (after every non-destructive section,
-          including the label sections) so the irreversible Delete All Data
-          action sits at the very bottom and isn't hit mid-scroll (#778). */}
-      <div className="mt-8 pt-6 border-t border-red-200 dark:border-red-900/50">
-        <h2 className="text-lg font-semibold text-red-600 dark:text-red-400 mb-1">{t("settings.dangerZone")}</h2>
-        <p className="text-sm text-gray-500 mb-4">
-          {t("settings.dangerZoneDesc")}
-        </p>
-
-        {!showDeleteConfirm ? (
-          <button
-            onClick={() => { setShowDeleteConfirm(true); setDeleteResult(null); }}
-            className="px-4 py-2 bg-red-50 dark:bg-red-900/50 text-red-600 dark:text-red-400 border border-red-300 dark:border-red-800 rounded text-sm hover:bg-red-100 dark:hover:bg-red-900 hover:text-red-700 dark:hover:text-red-300 transition-colors inline-flex items-center gap-2"
-          >
-            <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-              <path strokeLinecap="round" strokeLinejoin="round" d="m14.74 9-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 0 1-2.244 2.077H8.084a2.25 2.25 0 0 1-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 0 0-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 0 1 3.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 0 0-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 0 0-7.5 0" />
-            </svg>
-            {t("settings.deleteDatabase")}
-          </button>
-        ) : (
-          <div className="p-4 border border-red-300 dark:border-red-800 rounded-lg bg-red-50 dark:bg-red-950/30">
-            <p className="text-sm text-red-700 dark:text-red-300 mb-3">
-              {t("settings.deleteTypeConfirm")}
-            </p>
-            <div className="flex gap-2 items-center">
-              <input
-                type="text"
-                value={deleteInput}
-                onChange={(e) => setDeleteInput(e.target.value)}
-                onKeyDown={(e) => {
-                  if (e.key === "Escape") {
-                    setShowDeleteConfirm(false);
-                    setDeleteInput("");
-                  }
-                }}
-                placeholder={t("settings.deleteTypePlaceholder")}
-                className="w-40 px-3 py-1.5 border border-red-300 dark:border-red-800 rounded text-sm bg-transparent text-red-700 dark:text-red-200 placeholder-red-300 dark:placeholder-red-800 focus:outline-none focus:border-red-600"
-                autoFocus
-              />
-              <button
-                onClick={async () => {
-                  setDeleting(true);
-                  setDeleteResult(null);
-                  try {
-                    const res = await fetch("/api/snapshot/delete", { method: "DELETE" });
-                    const data = await res.json();
-                    if (res.ok) {
-                      setDeleteResult({
-                        ok: true,
-                        message: t("settings.deleteResult", { filaments: data.deleted.filaments, nozzles: data.deleted.nozzles, printers: data.deleted.printers }),
-                      });
-                    } else {
-                      setDeleteResult({ ok: false, message: data.error || t("settings.deleteFailed") });
-                    }
-                  } catch {
-                    setDeleteResult({ ok: false, message: t("common.serverError") });
-                  } finally {
-                    setDeleting(false);
-                    setShowDeleteConfirm(false);
-                    setDeleteInput("");
-                  }
-                }}
-                disabled={deleteInput.trim() !== "DELETE" || deleting}
-                className="px-4 py-1.5 bg-red-700 text-white rounded text-sm hover:bg-red-600 disabled:opacity-30 disabled:cursor-not-allowed transition-colors"
-              >
-                {deleting ? t("settings.deleting") : t("settings.confirmDelete")}
-              </button>
-              <button
-                onClick={() => { setShowDeleteConfirm(false); setDeleteInput(""); }}
-                className="px-3 py-1.5 text-gray-500 dark:text-gray-400 hover:text-gray-700 dark:hover:text-gray-200 text-sm transition-colors"
-              >
-                {t("common.cancel")}
-              </button>
-            </div>
-          </div>
-        )}
-
-        {deleteResult && (
-          <div className={`mt-3 text-sm px-3 py-2 rounded ${
-            deleteResult.ok
-              ? "bg-green-50 dark:bg-green-900/50 text-green-700 dark:text-green-300 border border-green-200 dark:border-green-800"
-              : "bg-red-50 dark:bg-red-900/50 text-red-700 dark:text-red-300 border border-red-200 dark:border-red-800"
-          }`}>
-            {deleteResult.message}
-          </div>
-        )}
-      </div>
-
-      <div className="mt-8 pt-6 border-t border-gray-200 dark:border-gray-800">
-        <p className="text-xs text-gray-600">
-          Filament DB v{process.env.APP_VERSION}
-        </p>
+      {/* Danger Zone — its own red tile → page */}
+      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4 mt-10">
+        <SettingsTile href="/settings/danger" title={t("settings.group.danger")} description={t("settings.group.danger.desc")} danger />
       </div>
     </main>
-  );
-}
-
-function ThemeSection() {
-  const { t } = useTranslation();
-  const { preference, setPreference } = useTheme();
-  const options: { code: ThemePreference; labelKey: string }[] = [
-    { code: "light", labelKey: "settings.theme.light" },
-    { code: "dark", labelKey: "settings.theme.dark" },
-    { code: "system", labelKey: "settings.theme.system" },
-  ];
-  return (
-    <div className="mt-8 pt-6 border-t border-gray-200 dark:border-gray-800">
-      <h2 className="text-lg font-semibold text-gray-900 dark:text-gray-200 mb-1">
-        {t("settings.theme.title")}
-      </h2>
-      <p className="text-sm text-gray-500 mb-4">{t("settings.theme.desc")}</p>
-      <div className="flex gap-2">
-        {options.map((opt) => (
-          <button
-            key={opt.code}
-            type="button"
-            onClick={() => setPreference(opt.code)}
-            className={`px-4 py-2 text-sm rounded border transition-colors ${
-              preference === opt.code
-                ? "border-blue-500 bg-blue-50 dark:bg-blue-600/20 text-blue-600 dark:text-blue-300"
-                : "border-gray-300 dark:border-gray-700 text-gray-600 dark:text-gray-400 hover:border-gray-400 dark:hover:border-gray-500 hover:text-gray-800 dark:hover:text-gray-300"
-            }`}
-          >
-            {t(opt.labelKey)}
-          </button>
-        ))}
-      </div>
-    </div>
   );
 }
