+++ src/components/LabelPrinterSettings.tsx
@@ -250,7 +250,7 @@ export default function LabelPrinterSettings() {
   if (!isElectron) return null;
 
   return (
-    <section className="mb-8">
+    <section>
       <h2 className="text-lg font-semibold text-gray-900 dark:text-gray-200 mb-1">
         {t("settings.labelPrinter")}
       </h2>
