+++ src/i18n/locales/de.json
@@ -1403,6 +1403,22 @@
   "common.retry": "Erneut versuchen",
   "settings.title": "Einstellungen",
   "settings.subtitle": "Konfigurieren Sie Ihre Filament DB-Instanz.",
+  "settings.back": "← Einstellungen",
+  "settings.group.manage": "Verwalten",
+  "settings.group.configure": "Konfigurieren",
+  "settings.group.ui": "UI-Einstellungen",
+  "settings.group.ui.desc": "Währung, Design und Sprache.",
+  "settings.group.network": "Netzwerkeinstellungen",
+  "settings.group.network.desc": "Datenbank-Verbindungsmodus und Freigabe im lokalen Netzwerk.",
+  "settings.group.data": "Sicherung & Daten",
+  "settings.group.data.desc": "Datenbank-Snapshot herunterladen oder wiederherstellen.",
+  "settings.group.ai": "KI",
+  "settings.group.ai.desc": "KI-Anbieter und API-Schlüssel für die Datenblatt-Extraktion.",
+  "settings.group.devices": "Geräte",
+  "settings.group.devices.desc": "NFC-Leser/-Tags und der Etikettendrucker.",
+  "settings.group.danger": "Gefahrenzone",
+  "settings.group.danger.desc": "Unwiderrufliche Aktionen, z. B. alle Daten löschen.",
+  "settings.network.desktopOnly": "Netzwerkeinstellungen sind nur in der Desktop-App verfügbar.",
   "settings.nozzles": "Düsen",
   "settings.nozzlesDesc": "Verwalten Sie Düsenprofile für Ihre Drucker.",
   "settings.printers": "Drucker",
