+++ src/i18n/locales/en.json
@@ -1403,6 +1403,22 @@
   "common.retry": "Retry",
   "settings.title": "Settings",
   "settings.subtitle": "Configure your Filament DB instance.",
+  "settings.back": "← Settings",
+  "settings.group.manage": "Manage",
+  "settings.group.configure": "Configure",
+  "settings.group.ui": "UI Settings",
+  "settings.group.ui.desc": "Currency, theme, and language.",
+  "settings.group.network": "Network Settings",
+  "settings.group.network.desc": "Database connection mode and sharing on your local network.",
+  "settings.group.data": "Backup & Data",
+  "settings.group.data.desc": "Download or restore a database snapshot.",
+  "settings.group.ai": "AI",
+  "settings.group.ai.desc": "AI provider and API key for technical-data-sheet extraction.",
+  "settings.group.devices": "Devices",
+  "settings.group.devices.desc": "NFC reader/tags and the label printer.",
+  "settings.group.danger": "Danger Zone",
+  "settings.group.danger.desc": "Irreversible actions, like deleting all data.",
+  "settings.network.desktopOnly": "Network settings are available in the desktop app only.",
   "settings.nozzles": "Nozzles",
   "settings.nozzlesDesc": "Manage nozzle profiles for your printers.",
   "settings.printers": "Printers",
